requirements.txt
torch
diffusers
transformers
accelerate
safetensors
huggingface_hub
gradio
Pillow
numpy
